+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- `/network/eras` duplicate era on Preview testnet
+
 ## [4.2.0] - 2025-09-24
 
 ### Added